Updated exception handling
Added a File.Exists(path) check because Icon.ExtractAssociatedIcon can throw a FileNotFoundException when the path is e.g. "C:\Windows\system32\ApplicationFrameHost.exe".
VibranceGUI is x86 so it runs under WOW64 on x64 systems and when it tries to access the system32 folder, the access will be redirected to %WinDir%\SysWOW64 but there isn't any ApplicationFrameHost.exe.

Removed exception logging for closing processes because it's simply unnecessary and slowly creates useless data over time.
It's very unlikely that they will happen when not shutting down the PC.